feat: tab component 수정 (#16)
